+++ test/unit/sonarqube_code_coverage.sh
@@ -10,6 +10,7 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/create_header.py
 coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/data_out.py
+coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/get_dumps.py
 coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/help_message.py
 coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/rename_repo.py
 coverage run -a --source=elastic_db_repo test/unit/elastic_db_repo/rename.py